# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CarePulse AI health triage assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface RegisterInput {
  /**
     * @minLength 2
     * @maxLength 255
     */
  name: string;
  email: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type UrgencyLevel = typeof UrgencyLevel[keyof typeof UrgencyLevel];


export const UrgencyLevel = {
  'Self-Care': 'Self-Care',
  Routine_Consultation: 'Routine Consultation',
  Urgent_Care: 'Urgent Care',
  EMERGENCY_IMMEDIATE_CARE: 'EMERGENCY_IMMEDIATE_CARE',
} as const;

export type Likelihood = typeof Likelihood[keyof typeof Likelihood];


export const Likelihood = {
  Low: 'Low',
  Moderate: 'Moderate',
  High: 'High',
} as const;

export interface PotentialCause {
  name: string;
  likelihood: Likelihood;
  description: string;
}

export interface TriageAnalysis {
  urgencyLevel: UrgencyLevel;
  disclaimer: string;
  summary: string;
  potentialCauses: PotentialCause[];
  recommendedActions: string[];
  questionsForDoctor: string[];
  redFlagWarnings: string[];
}

export interface AssessmentInput {
  /**
     * @minimum 0
     * @maximum 120
     */
  age: number;
  /**
     * @minLength 1
     * @maxLength 50
     */
  biologicalSex: string;
  /**
     * @minLength 3
     * @maxLength 5000
     */
  primarySymptoms: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  duration: string;
  /**
     * @minimum 1
     * @maximum 10
     */
  severity: number;
  /** @maxLength 3000 */
  chronicConditions?: string;
  /** @maxLength 3000 */
  currentMedications?: string;
}

export interface Assessment {
  id: number;
  age: number;
  biologicalSex: string;
  primarySymptoms: string;
  duration: string;
  severity: number;
  /** @nullable */
  chronicConditions: string | null;
  /** @nullable */
  currentMedications: string | null;
  urgencyLevel: UrgencyLevel;
  aiAnalysis: TriageAnalysis;
  createdAt: string;
}

export type DashboardSummaryUrgencyCounts = {[key: string]: number};

export interface DashboardSummary {
  totalAssessments: number;
  thisMonth: number;
  latestAssessment: Assessment | null;
  urgencyCounts: DashboardSummaryUrgencyCounts;
}

export type ListAssessmentsParams = {
search?: string;
urgency?: UrgencyLevel;
};