# lib/api-zod/src/generated/types/assessment.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CarePulse AI health triage assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { TriageAnalysis } from './triageAnalysis';
import type { UrgencyLevel } from './urgencyLevel';

export interface Assessment {
  id: number;
  age: number;
  biologicalSex: string;
  primarySymptoms: string;
  duration: string;
  severity: number;
  /** @nullable */
  chronicConditions: string | null;
  /** @nullable */
  currentMedications: string | null;
  urgencyLevel: UrgencyLevel;
  aiAnalysis: TriageAnalysis;
  createdAt: Date;
}