# lib/api-zod/src/generated/types/assessmentInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CarePulse AI health triage assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface AssessmentInput {
  /**
     * @minimum 0
     * @maximum 120
     */
  age: number;
  /**
     * @minLength 1
     * @maxLength 50
     */
  biologicalSex: string;
  /**
     * @minLength 3
     * @maxLength 5000
     */
  primarySymptoms: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  duration: string;
  /**
     * @minimum 1
     * @maximum 10
     */
  severity: number;
  /** @maxLength 3000 */
  chronicConditions?: string;
  /** @maxLength 3000 */
  currentMedications?: string;
}