# lib/api-zod/src/generated/types/authResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CarePulse AI health triage assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface AuthResponse {
  token: string;
  user: User;
}