# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CarePulse AI health triage assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { Assessment } from './assessment';
import type { DashboardSummaryUrgencyCounts } from './dashboardSummaryUrgencyCounts';

export interface DashboardSummary {
  totalAssessments: number;
  thisMonth: number;
  latestAssessment: Assessment | null;
  urgencyCounts: DashboardSummaryUrgencyCounts;
}