# lib/api-zod/src/generated/types/likelihood.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CarePulse AI health triage assistant API
 * OpenAPI spec version: 0.1.0
 */

export type Likelihood = typeof Likelihood[keyof typeof Likelihood];


export const Likelihood = {
  Low: 'Low',
  Moderate: 'Moderate',
  High: 'High',
} as const;