# lib/api-zod/src/generated/types/potentialCause.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CarePulse AI health triage assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { Likelihood } from './likelihood';

export interface PotentialCause {
  name: string;
  likelihood: Likelihood;
  description: string;
}