# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CarePulse AI health triage assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  /**
     * @minLength 2
     * @maxLength 255
     */
  name: string;
  email: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}