# lib/api-zod/src/generated/types/triageAnalysis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CarePulse AI health triage assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { PotentialCause } from './potentialCause';
import type { UrgencyLevel } from './urgencyLevel';

export interface TriageAnalysis {
  urgencyLevel: UrgencyLevel;
  disclaimer: string;
  summary: string;
  potentialCauses: PotentialCause[];
  recommendedActions: string[];
  questionsForDoctor: string[];
  redFlagWarnings: string[];
}