# lib/api-zod/src/generated/types/urgencyLevel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CarePulse AI health triage assistant API
 * OpenAPI spec version: 0.1.0
 */

export type UrgencyLevel = typeof UrgencyLevel[keyof typeof UrgencyLevel];


export const UrgencyLevel = {
  'Self-Care': 'Self-Care',
  Routine_Consultation: 'Routine Consultation',
  Urgent_Care: 'Urgent Care',
  EMERGENCY_IMMEDIATE_CARE: 'EMERGENCY_IMMEDIATE_CARE',
} as const;